requirements.txt
fastapi>=0.110.0
uvicorn[standard]==0.24.0
aiohttp==3.9.1
apscheduler==3.10.4
python-multipart>=0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic>=2.6.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-whois==0.8.0
tldextract==5.1.1